Add requirements.txt and improve code quality
Add `requirements.txt` file and update dependencies.

* **Update `password_list.txt`**: Remove comment lines and ensure the file contains only passwords.
* **Update `readme.md`**: Include installation instructions for dependencies using `pip install -r requirements.txt`.
* **Rename and refactor `ZIPPY.py` to `zippy/main.py`**: Break down the large file into smaller modules, add logging, improve error handling, and add docstrings and inline comments.
* **Add `zippy/extract.py`**: Add functions related to extracting archives.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/John0n1/ZIPPY?shareId=XXXX-XXXX-XXXX-XXXX).
