+++ password_list.txt
@@ -1,5 +1,3 @@
-// Your Password List e.g:
 123qwe
 1243ed
 123456789
-//
